fix merge issues

+++ src/app/snap.h
@@ -13,6 +13,6 @@ void func_80339E1C(VehicleData*);
 s32 func_8033E3A8(s32);
 s32 func_8033F62C(void);
 s32 snapGetPhotoCount(void);
-s16 func_80338614(void);
+s32 func_80338614(void);
 
 #endif // APP_SNAP_H
